siirretty asioita tiedostoihin ja kansioihin
